+++ compose/src/commonMain/kotlin/com/napos/stepper/compose/components/StepperContentType.kt
@@ -1,0 +1,6 @@
+package com.napos.stepper.compose.components
+
+public enum class StepperContentType {
+    Preview,
+    Step,
+}