+++ example/src/commonMain/kotlin/com/napos/stepper/example/steps/second/SecondScreen.kt
@@ -32,4 +32,13 @@ class SecondScreen(
         )
     }
 
+    @Composable
+    override fun preview() {
+        milestone.data?.let {
+            Text(it.toString())
+        } ?: run {
+            Text("No data")
+        }
+    }
+
 }