+++ example/src/commonMain/kotlin/com/napos/stepper/example/steps/third/ThirdScreen.kt
@@ -30,4 +30,13 @@ class ThirdScreen(
             label = { Text("Third value") }
         )
     }
+
+    @Composable
+    override fun preview() {
+        milestone.data?.let {
+            Text(it.toString())
+        } ?: run {
+            Text("No data")
+        }
+    }
 }